# Make MathOptInterface.jl a weak dependency