--- a/Project.toml
+++ b/Project.toml
@@ -10,10 +10,10 @@ LineSearches = "d3d80556-e9d4-5f37-9878-2ab0fcc64255"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 NLSolversBase = "d41bc354-129a-5804-8e4c-c37616107c6c"
 NaNMath = "77ba4419-2d1f-58cd-9bb1-8ffee604a2e3"
+PackageExtensionCompat = "65ce6f38-6b18-4e1d-a461-8949797d7930"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 PositiveFactorizations = "85a6dd25-e78a-55b7-8502-1745935b8125"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
@@ -33,11 +33,11 @@ MathOptInterface = "1.17"
 NLSolversBase = "7.8.0"
 NaNMath = "0.3.2, 1"
 OptimTestProblems = "2.0.3"
+PackageExtensionCompat = "1"
 Parameters = "0.10, 0.11, 0.12"
 PositiveFactorizations = "0.2.2"
 Printf = "<0.0.1, 1.6"
 Random = "<0.0.1, 1.6"
-Requires = "1"
 SparseArrays = "<0.0.1, 1.6"
 StatsBase = "0.29, 0.30, 0.31, 0.32, 0.33, 0.34"
 Test = "<0.0.1, 1.6"
@@ -56,16 +56,4 @@ StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = [
-    "Test",
-    "Distributions",
-    "Measurements",
-    "OptimTestProblems",
-    "Random",
-    "RecursiveArrayTools",
-    "StableRNGs",
-    "LineSearches",
-    "NLSolversBase",
-    "PositiveFactorizations",
-    "MathOptInterface",
-]
+test = ["Test", "Distributions", "Measurements", "OptimTestProblems", "Random", "RecursiveArrayTools", "StableRNGs", "LineSearches", "NLSolversBase", "PositiveFactorizations"]