--- a/src/Optim.jl
+++ b/src/Optim.jl
@@ -34,6 +34,7 @@ using Printf                 # For printing, maybe look into other options
 using FillArrays             # For handling scalar bounds in Fminbox
 
 #using Compat                 # for compatibility across multiple julia versions
+using PackageExtensionCompat # For retrocompatibility on package extensions
 
 # for extensions of functions defined in Base.
 import Base: length, push!, show, getindex, setindex!, maximum, minimum
@@ -220,11 +221,10 @@ include("multivariate/solvers/constrained/ipnewton/utilities/trace.jl")
 # Maximization convenience wrapper
 include("maximize.jl")
 
-# MOI Related utilities
-"Constructor for a MOI compatible optimizer"
-function moi_optimizer end
-
-using PackageExtensionCompat
+# Pre assign Optimizer, to be overloaded in OptimMOIExt.
+if VERSION >= v"1.9"
+  global Optimizer
+end
 
 function __init__()
     @require_extensions