MOI adds a very large loading time to Optim.jl although it is not always necessary see : https://github.com/jump-dev/MathOptInterface.jl/issues/2442

This PR makes MOI.jl a weak dependency, this is the difference of time for precompiling the package:

`master`:
```
julia> @time Pkg.precompile("Optim")
Precompiling Optim
  2 dependencies successfully precompiled in 112 seconds. 53 already precompiled.
113.622423 seconds (776.52 k allocations: 56.027 MiB, 0.02% gc time, 1.25% compilation time: 92% of which was recompilation)
julia> @time using Optim
  3.414563 seconds (3.74 M allocations: 231.737 MiB, 3.75% gc time, 0.28% compilation time)
```

This branch
```
julia> @time Pkg.precompile("Optim")
  4.763922 seconds (2.90 M allocations: 213.363 MiB, 1.90% gc time, 91.22% compilation time: 92% of which was recompilation)

julia> @time using Optim
  0.401095 seconds (451.49 k allocations: 30.527 MiB, 6.22% gc time, 2.45% compilation time)
```

For retro-compatibility I used `Requires.jl` as suggested in the [docs](https://pkgdocs.julialang.org/v1/creating-packages/#Requires.jl).

One important change is that I could not create a reference to `Optim.Optimizer <: MOI.AbstractOptimizer` in `Optim`, so I resorted to make a constructor function `moi_optimizer()` that is then defined in `OptimMOIExt.jl`.