@theogf 2024-03-10
@MilesCranmer I followed your suggestion and replace Requires by PackageExtensionCompat