@blegat 2024-03-11
You can keep the name `Optim.Optimizer` see for instance what we did for NLopt