@theogf 2024-03-11
@MilesCranmer The solution in NLopt would mean switching back to the `Requires` solution as it needs more fine tuning, I guess keeping the consistency with the previous version is more important?
Note that under the hood, `PackageExtensionCompat` is just using `Requires`.