@MilesCranmer 2024-03-11
> Note that under the hood, PackageExtensionCompat is just using Requires.

Yeah this is why I suggested removing the `Requires` line, since you aren't using it here. Therefore on Julia 1.9+, Requires.jl won't be loaded unnecessarily.