@MilesCranmer 2024-03-11
What @odow is suggesting is not incompatible with PackageExtensionCompat; you just need to define `Optimizer` in the main codebase and then overload it in the extension.