@blegat 2024-03-13
Thanks! Yes, it's a bit delicate to get this working for both Julia v1.6 and v1.10 ^^