@MilesCranmer 2024-03-13
> It allowed me to define Optim.Optimizer as a const for v1.8-

Then I would do the following:

```julia
if VERSION >= v"1.9.0-DEV.0"
    @eval const Optimizer = OptimMOIExt.Optimizer
else
    @eval Optimizer = OptimMOIExt.Optimizer
end
```