@theogf 2024-03-14
> > It allowed me to define Optim.Optimizer as a const for v1.8-
> 
> Then I would do the following:
> 
> ```julia
> if VERSION >= v"1.9.0-DEV.0"
>     @eval const Optimizer = OptimMOIExt.Optimizer
> else
>     @eval Optimizer = OptimMOIExt.Optimizer
> end
> ```

I'd rather avoid using `@eval` if possible (could not find another solution for <v1.9).