@MilesCranmer 2024-03-14
What's wrong with an `@eval`? It's all static expressions here