@MilesCranmer 2024-03-15
Oh, weird. What is the purpose of the `[weakdeps]` in that case?