@theogf 2024-03-16
I am not sure what the conclusion is here :sweat_smile: 

Should I implement the solution proposed by @blegat and @ericphanson ?